Refactor ChatHistoryFragment boundary to use ChatViewModel
- Added TeamsRepository and VoicesRepository to ChatViewModel's constructor.
- Moved loadShareTargets method to ChatViewModel.
- Moved shareChatToVoices logic to ChatViewModel.
- Updated ChatHistoryFragment to delegate share target and news creation logic to ChatViewModel.
- Updated unit tests for ChatViewModel.
